+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - New module `raw::types::direct`, containing raw `Deserialize`able types used with the Direct
   Message API
 - New function `Response::try_map`, to change the type of a `Response` using a fallible conversion
+- New function `Response::into`, to convert a `Response` using the `Into` trait
 - `Response` now implements `IntoIterator` if its contained type does
   - A new type `ResponseIter` has been introduced to contain this iterator